+++ test/object_extension_test.dart
@@ -1,6 +1,8 @@
 import 'package:dart_extensionz/dart_extensionz.dart';
 import 'package:test/test.dart';
 
+import 'enum_extension_test.dart';
+
 void main() {
   Object? o0;
   const Object? o1 = null;
@@ -39,5 +41,27 @@ void main() {
     expect(m1.isBlank, true);
     expect(m2.isBlank, true);
     expect(m3.isBlank, false);
+  });
+
+  test('types', () {
+    expect(true.isBool, true);
+    expect(4.0.isDouble, true);
+    expect(TestEnum.one.isEnum, true);
+    expect(1.isInt, true);
+    expect(<String>['test'].isList, true);
+    expect(<String, String>{'key': 'value'}.isMap, true);
+    expect(10.isNum, true);
+    expect(<String>{'test'}.isSet, true);
+    expect('test'.isString, true);
+
+    expect(null.isBool, false);
+    expect(null.isDouble, false);
+    expect(null.isEnum, false);
+    expect(null.isInt, false);
+    expect(null.isList, false);
+    expect(null.isMap, false);
+    expect(null.isNum, false);
+    expect(null.isSet, false);
+    expect(null.isString, false);
   });
 }